refactor: Provider 와 ProviderImpl 파일 분리
